Add staged trials, and total counts.
We now want to show trials where letters have been sent to
journals, but the journals haven't yet published them. For these
trials, we only want to show the fact that the trial exists, and
the date the letter was sent.

We also want a summary paragraph with counts of total trials.